requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
argcomplete==2.0.0
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.0.4
colorama==0.4.4
crytic-compile==0.2.2
cytoolz==0.11.2
dataclasses-json==0.5.7
distlib==0.3.4
enum34==1.1.10
eth-hash==0.3.2
eth-typing==3.0.0
eth-utils==2.0.0
exceptiongroup==1.1.1
filelock==3.6.0
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.2
greenlet==2.0.2
h11==0.14.0
huggingface-hub==0.14.1
idna==3.4
intervaltree==3.1.0
Jinja2==3.0.3
langchain==0.0.162
lxml==4.9.2
MarkupSafe==2.1.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.1
numexpr==2.8.4
numpy==1.22.2
openai==0.27.6
openapi-schema-pydantic==1.2.4
outcome==1.2.0
packaging==21.3
pandas==1.4.1
pefile==2021.9.3
Pillow==9.0.1
pipx==1.0.0
platformdirs==2.5.1
ply==3.11
prettytable==3.2.0
progressbar2==4.0.0
protobuf==3.20.0
psycopg2==2.9.6
py2exe==0.11.1.0
pycparser==2.21
pydantic==1.10.7
pyevmasm==0.2.3
pyparsing==3.0.7
pyperclip==1.8.2
pysha3==1.0.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-utils==3.1.0
pytz==2021.3
PyYAML==6.0
requests==2.30.0
rlp==3.0.0
selenium==4.9.1
six==1.16.0
slither-analyzer==0.8.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.12
tenacity==8.2.2
toolz==0.11.2
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==2.0.2
userpath==1.8.0
virtualenv==20.13.2
wasm==1.2
wcwidth==0.2.5
wsproto==1.2.0
yarl==1.9.2
z3-solver==4.8.15.0
